requirement.txt
alembic==1.4.3
appdirs==1.4.4
black==20.8b1
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
pathspec==0.8.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
regex==2020.9.27
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
